requirements.txt
# E-Commerce Flask Application Requirements
# Version: 1.0.0
# Python Version: 3.8+

# Core Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Moment==1.0.5
WTForms==3.0.1
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Environment Configuration
python-dotenv==1.0.0

# Image Processing
Pillow==10.0.1

# Password Hashing
bcrypt==4.0.1

# Form Validation and Security
email-validator==2.0.0

# Testing Framework
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-html==4.1.1
coverage==7.3.2

# Selenium WebDriver for UI Testing
selenium==4.15.2
webdriver-manager==4.0.1

# Production Server
gunicorn==21.2.0

# Test Data Generation
Faker==19.6.2
factory-boy==3.3.0

# Additional Dependencies
Flask-Mail==0.9.1
python-dateutil==2.8.2
requests==2.31.0
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.6.3
alembic==1.12.0